// Web/src/app/models/CustomerResponse.ts
export class CustomerResponse {
  customerId?: number;
  customerKey?: string;
  customerName?: string;
  supported?: boolean;
  supportContractExpirationDate?: string;
  address1?: string;
  address2?: string;
  city?: string;
  state?: string;
  zip?: string;
  country?: string;
  phone1?: string;
  phone2?: string;
  fax?: string;
  email?: string;
  email2?: string;
  website?: string;
  contactFirstName?: string;
  contactLastName?: string;
  contactPhone1?: string;
  contactPhone2?: string;
  contactFax?: string;
  contactEmail?: string;
  notes?: string;
  databaseStatus?: string;
  lastActivity?: string;
  databaseStatusChanged?: string;
  active?: boolean;
  deleted?: boolean;
  modifiedOn?: string;
  modifiedBy?: string;
  createdOn?: string;
  createdBy?: string;
  apikey?: string;
  offset?: string;
  expirationDate?: string;
  isDemo?: boolean;
  logInCount?: number;
  isQbo?: boolean;
  openIdqbo?: string;
  cancellationNumber?: string;
  confirmRenewal?: boolean;
  enterpriseId?: number;
  numberOfSites?: number;
}